domain/analyze/ai_agent/requirements.txt
# AI/ML Core
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
anthropic>=0.7.0
openai>=1.0.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.8.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Web Interface
streamlit>=1.28.0
flask>=2.3.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# File Processing
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# System
psutil>=5.9.0
pathlib>=1.0.1